requirements.txt
azure-core==1.30.2
azure-identity==1.17.1
azure-storage-blob==12.21.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.7
isodate==0.6.1
msal==1.30.0
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
